Make HasDROP and HasBUFF usable directly on UpdateFAR/CreateFAR/DownlinkDataReport IEs
Having 2 different behaviours depending on the method is not expected.
Having only one function is by the way a curious choice since BUFF/DROP
doesn't have exactly the same sementic for Data Status and Apply Action…
In my opinion, helper functions should always be separated by IE and
return a `(bool, err)` with `ErrInvalidType` when wrongly used. This
would be a breaking change, so I will not doing it in this PR, and would
like to have a second opinion before any implementation.